add observer unit
