+++ src/tests/integration/dependency-injection.test.ts
@@ -1,5 +1,4 @@
 import { Delegate, Event } from '../../exports/exports'
-import { assert } from 'console'
 import { test } from '../test.test'
 
 class Consumer {
@@ -7,20 +6,16 @@ class Consumer {
 		onTest.add(this.consume)
 	}
 
-	consume = (is: boolean) => {
-		this.is = is
-	}
+	consume = (is: boolean) => (this.is = is)
 
 	is = false
 }
 
 function dependencyInjection() {
-	const delegate = new Delegate<[boolean]>(),
+	const delegate = new Delegate([true]),
 		event = new Event(delegate)
 
 	const consumer = new Consumer(event)
-
-	delegate(true)
 
 	return consumer.is
 }
