--- src/tests/integration/observer-pattern.test.ts
@@ -1,25 +1,0 @@
-import { Delegate } from '../../exports/exports'
-import { assert } from 'console'
-
-class Subject<T> extends Delegate<[T]> {
-	subscribe({ notify }: { notify(t: T): void }) {
-		this.add(notify)
-	}
-}
-
-async function test() {
-	const subject = new Subject<boolean>([false]),
-		observer = { notify: (message: boolean) => (is = message) }
-
-	let is = false
-	subject.subscribe(observer)
-	subject(true)
-
-	return is
-}
-
-try {
-	assert(test(), '❌ observer-pattern')
-} catch (e) {
-	assert(false, '❌ observer-pattern: ', e)
-}
